+++ TradeCommon/Essentials/Instruments/SimpleSecurity.cs
@@ -1,6 +1,6 @@
 ﻿namespace TradeCommon.Essentials.Instruments;
 
-public record SimpleSecurity(int Id, string Code, string Name, string Exchange, string Type)
+public record SimpleSecurity(long Id, string Code, string Name, string Exchange, string Type)
 {
     public SimpleSecurity(Security security) : this(security.Id, security.Code, security.Name, security.Exchange, security.Type)
     {
