+++ TradeDataCore/Importing/Binance/HistoricalPriceReader.cs
@@ -10,9 +10,9 @@ public class HistoricalPriceReader : IHistoricalPriceReader
 {
     private static readonly ILog _log = Logger.New();
 
-    public async Task<Dictionary<int, List<OhlcPrice>>?> ReadPrices(List<Security> securities, DateTime start, DateTime end, IntervalType intervalType)
+    public async Task<Dictionary<long, List<OhlcPrice>>?> ReadPrices(List<Security> securities, DateTime start, DateTime end, IntervalType intervalType)
     {
-        var results = new Dictionary<int, List<OhlcPrice>>();
+        var results = new Dictionary<long, List<OhlcPrice>>();
         await Parallel.ForEachAsync(securities, async (security, ct) =>
         {
             var prices = await ReadPrices(security, start, end, intervalType);
