+++ TradeLogicCore/Program.cs
@@ -416,7 +416,7 @@ public class Program
         }
         await dataService.UnsubscribeOhlc(security, interval);
 
-        void OnNewOhlc(int securityId, OhlcPrice price, IntervalType interval, bool isComplete)
+        void OnNewOhlc(long securityId, OhlcPrice price, IntervalType interval, bool isComplete)
         {
             printCount++;
             if (security != null && security.Id != securityId)
